Move `deployments` to `deployment-list` and flatten directory structure
